training-loop/requirements.txt
confluent-kafka==2.3.0
orjson==3.9.15
psycopg2-binary==2.9.9
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
